Version Shantall 1.1 (Trte de mejorar catalgo y no es funcional aun)

+++ app/src/main/java/com/example/sanitaapp/model/Planta.kt
@@ -4,6 +4,6 @@ data class Planta(
     val id: Int,
     val nombre: String,
     val imagen: String,
-    val precio: Double, // Precio
+    val precio: Int, // Precio
     val categoria: Int, // Nueva campo para la categoría
 )
